check end events and connectionReleased
