+++ okhttp/src/main/java/okhttp3/EventListener.java
@@ -119,7 +119,7 @@ public abstract class EventListener {
    * connection attempt.
    */
   public void connectEnd(Call call, InetSocketAddress inetSocketAddress,
-      @Nullable Protocol protocol, @Nullable Throwable throwable) {
+      @Nullable Protocol protocol, @Nullable Proxy proxy, @Nullable Throwable throwable) {
   }
 
   /**
