+++ okhttp/src/main/java/okhttp3/internal/connection/StreamAllocation.java
@@ -287,6 +287,9 @@ public final class StreamAllocation {
       this.codec = null;
     }
     if (released) {
+      if (connection != null) {
+        eventListener.connectionReleased(call, connection);
+      }
       this.released = true;
     }
     Socket socket = null;
